+++ tests/compat.rs
@@ -51,8 +51,7 @@ mod pgp {
                 cargo_file.to_str().unwrap(),
                 FileOptions::new("/etc/foobar/hugo/bazz.toml")
                     .mode(0o100_777)
-                    .is_config()
-                    .is_no_replace(),
+                    .is_config_noreplace(),
             )?
             .with_file(
                 cargo_file.to_str().unwrap(),