+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Dependency::rpmlib()` now inserts the `rpmlib()` portion automatically, only the feature name itself should
   be provided in the string passed as the name argument.
+- `FileOptions::is_no_replace()` is now `FileOptions::is_config_noreplace()` to reflect the fact that the noreplace
+  flag is only applicable to config files, and have more similar usage relative to `%config(noreplace)`
 
 ### Added